+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/resourcecontainer/RcConstants.kt
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Orature.  If not, see <https://www.gnu.org/licenses/>.
  */
-package org.wycliffeassociates.otter.common.domain.resourcecontainer.projectimportexport
+package org.wycliffeassociates.otter.common.domain.resourcecontainer
 
 internal object RcConstants {
     const val MEDIA_DIR = "content"
